+++ Client/N3Base/N3GESnow.cpp
@@ -46,7 +46,7 @@ void CN3GESnow::Tick()
 	{
 		if(m_fFadeTime > 0 && m_fFadeTimeCur < m_fFadeTime) // Fade½Ã°£¶«¿¡ °Ç³Ê¶Ù°í ÂïÀ» ¾ç °áÁ¤..
 		{
-			iActiveCount = iCount * m_fFadeTimeCur / m_fFadeTime;
+			iActiveCount = (int)(iCount * m_fFadeTimeCur / m_fFadeTime);
 			if(iActiveCount > iCount) iActiveCount = iCount;
 			else if(iActiveCount < 0) iActiveCount = 0;
 		}
@@ -55,7 +55,7 @@ void CN3GESnow::Tick()
 	{
 		if(m_fFadeTime > 0 && m_fFadeTimeCur < m_fFadeTime) // Fade½Ã°£¶«¿¡ °Ç³Ê¶Ù°í ÂïÀ» ¾ç °áÁ¤..
 		{
-			iActiveCount = iCount * (1.0f - m_fFadeTimeCur / m_fFadeTime);
+			iActiveCount = (int)(iCount * (1.0f - m_fFadeTimeCur / m_fFadeTime));
 			if(iActiveCount > iCount) iActiveCount = iCount;
 			else if(iActiveCount < 0) iActiveCount = 0;
 		}
@@ -145,7 +145,7 @@ void CN3GESnow::Render(__Vector3& vPos)
 	{
 		if(m_fFadeTime > 0 && m_fFadeTimeCur < m_fFadeTime) // Fade½Ã°£¶«¿¡ °Ç³Ê¶Ù°í ÂïÀ» ¾ç °áÁ¤..
 		{
-			iActiveCount = iCount * m_fFadeTimeCur / m_fFadeTime;
+			iActiveCount = (int)(iCount * m_fFadeTimeCur / m_fFadeTime);
 			if(iActiveCount > iCount) iActiveCount = iCount;
 			else if(iActiveCount < 0) iActiveCount = 0;
 		}
@@ -154,7 +154,7 @@ void CN3GESnow::Render(__Vector3& vPos)
 	{
 		if(m_fFadeTime > 0 && m_fFadeTimeCur < m_fFadeTime) // Fade½Ã°£¶«¿¡ °Ç³Ê¶Ù°í ÂïÀ» ¾ç °áÁ¤..
 		{
-			iActiveCount = iCount * (1.0f - m_fFadeTimeCur / m_fFadeTime);
+			iActiveCount = (int)(iCount * (1.0f - m_fFadeTimeCur / m_fFadeTime));
 			if(iActiveCount > iCount) iActiveCount = iCount;
 			else if(iActiveCount < 0) iActiveCount = 0;
 		}
