+++ Client/N3Base/N3SndObj.cpp
@@ -429,7 +429,7 @@ void CN3SndObj::Tick()
 		else
 		{
 			int vol = 0;
-			if(m_fFadeInTime>0.0f) vol = ((m_fTmpSecPerFrm/m_fFadeInTime)*(float)m_iMaxVolume);
+			if(m_fFadeInTime>0.0f) vol = (int)((m_fTmpSecPerFrm/m_fFadeInTime)*(float)m_iMaxVolume);
 			SetVolume(vol);
 		}
 	}
@@ -451,7 +451,7 @@ void CN3SndObj::Tick()
 		{
 			//볼륨 점점 작게....
 			int vol = 0;
-			if(m_fFadeOutTime>0.0f)  vol = (((m_fFadeOutTime - m_fTmpSecPerFrm)/m_fFadeOutTime)*(float)m_iMaxVolume);
+			if(m_fFadeOutTime>0.0f)  vol = (int)(((m_fFadeOutTime - m_fTmpSecPerFrm)/m_fFadeOutTime)*(float)m_iMaxVolume);
 			SetVolume(vol);
 		}
 	}
